refactor folder pages
