+++ src/app/pages/edit-employee/edit-employee.component.ts
@@ -1,33 +1,40 @@
 import { HttpErrorResponse } from '@angular/common/http';
 import { Component, OnInit, Optional } from '@angular/core';
-import { Employee } from '../employee';
-import { EmployeeService } from '../employee.service';
+import { Employee } from '../../model/employee';
+import { EmployeeService } from '../../service/employee.service';
 import { ModalDismissReasons, NgbModal } from '@ng-bootstrap/ng-bootstrap';
-import { NgForm, ReactiveFormsModule , } from '@angular/forms';
-
-
+import { NgForm } from '@angular/forms';
+import { ActivatedRoute } from '@angular/router';
 
 @Component({
-  selector: 'app-add-employee',
-  templateUrl: './add-employee.component.html',
+  selector: 'app-edit-employee',
+ templateUrl: `./edit-employee.component.html`,
   styles: [
   ]
 })
-export class AddEmployeeComponent implements OnInit {
-
-
+export class EditEmployeeComponent implements OnInit {
+  
   title = 'FINAL';
   public employees!:Employee[];
   closeResult!: string;
   public editEmployee!: Employee;
   public deleteEmployee!: Employee;
+  public idForUpdate!: number;
+
 
 
   constructor(private employeeService : EmployeeService,
-              private modalService: NgbModal) { }
+              private route:ActivatedRoute) { }
 
   ngOnInit(){
+
     this.getEmployees();
+
+    this.route.params.subscribe(params =>{
+      this.idForUpdate = params['id'];
+    });
+    this.getEmployeeById();
+
   }
 
   public getEmployees() :void{
@@ -40,6 +47,22 @@ export class AddEmployeeComponent implements OnInit {
       }
       );
   }
+
+
+  public getEmployeeById():void {
+    this.employeeService.getEmployeeById(this.idForUpdate).subscribe(
+      (response :Employee) => {
+        this.editEmployee = response;
+        console.log("AJUNGE AICI  GET EMPLOYYE BY ID");
+      },
+      (error :HttpErrorResponse) => {
+        alert(error.message);
+      }
+    )
+  }
+
+
+
 
   public onOpenModal(employee: Employee|null, mode: string) :void{
 
@@ -76,27 +99,29 @@ export class AddEmployeeComponent implements OnInit {
     button.click();}
 
   
-  open(content:any) {
-    this.modalService.open(content, {ariaLabelledBy: 'modal-basic-title'}).result.then((result) => {
-      this.closeResult = `Closed with: ${result}`;
-    }, (reason) => {
-      this.closeResult = `Dismissed ${this.getDismissReason(reason)}`;
-    });
-  }
   
-  private getDismissReason(reason: any): string {
-    if (reason === ModalDismissReasons.ESC) {
-      return 'by pressing ESC';
-    } else if (reason === ModalDismissReasons.BACKDROP_CLICK) {
-      return 'by clicking on a backdrop';
-    } else {
-      return `with: ${reason}`;
-    }
+
+
+
+
+  public onUpdateEmloyee(employee : any): void {
+
+    console.log("sadasdas");
+    console.log(this.idForUpdate);
+    console.log(employee);
+    this.employeeService.updateEmployee(this.idForUpdate,employee).subscribe(
+      (response: Employee) => {
+        console.log(response);
+        console.log("AICI0");
+        this.getEmployees();
+      
+      },
+      (error: HttpErrorResponse) => {
+        alert(error.message);
+      
+      }
+    );
   }
-
-
-
-
 
   public onAddEmloyee(addForm: NgForm): void {
     document.getElementById('add-employee-form')!.click();
@@ -125,5 +150,6 @@ export class AddEmployeeComponent implements OnInit {
       }
     );
   }
+
 
 }
