+++ src/app/pages/overview/overview.component.ts
@@ -1,40 +1,32 @@
+
 import { HttpErrorResponse } from '@angular/common/http';
 import { Component, OnInit, Optional } from '@angular/core';
-import { Employee } from '../employee';
-import { EmployeeService } from '../employee.service';
+
+
 import { ModalDismissReasons, NgbModal } from '@ng-bootstrap/ng-bootstrap';
 import { NgForm } from '@angular/forms';
-import { ActivatedRoute } from '@angular/router';
+import { EmployeeService } from 'src/app/service/employee.service';
+import { Employee } from 'src/app/model/employee';
 
 @Component({
-  selector: 'app-edit-employee',
- templateUrl: `./edit-employee.component.html`,
+  selector: 'app-overview',
+  templateUrl: `./overview.component.html`,
   styles: [
   ]
 })
-export class EditEmployeeComponent implements OnInit {
-  
+export class OverviewComponent implements OnInit {
   title = 'FINAL';
   public employees!:Employee[];
   closeResult!: string;
   public editEmployee!: Employee;
   public deleteEmployee!: Employee;
-  public idForUpdate!: number;
-
 
 
   constructor(private employeeService : EmployeeService,
-              private route:ActivatedRoute) { }
+              private modalService: NgbModal) { }
 
   ngOnInit(){
-
     this.getEmployees();
-
-    this.route.params.subscribe(params =>{
-      this.idForUpdate = params['id'];
-    });
-    this.getEmployeeById();
-
   }
 
   public getEmployees() :void{
@@ -47,22 +39,6 @@ export class EditEmployeeComponent implements OnInit {
       }
       );
   }
-
-
-  public getEmployeeById():void {
-    this.employeeService.getEmployeeById(this.idForUpdate).subscribe(
-      (response :Employee) => {
-        this.editEmployee = response;
-        console.log("AJUNGE AICI  GET EMPLOYYE BY ID");
-      },
-      (error :HttpErrorResponse) => {
-        alert(error.message);
-      }
-    )
-  }
-
-
-
 
   public onOpenModal(employee: Employee|null, mode: string) :void{
 
@@ -99,29 +75,25 @@ export class EditEmployeeComponent implements OnInit {
     button.click();}
 
   
-  
-
-
-
-
-  public onUpdateEmloyee(employee : any): void {
-
-    console.log("sadasdas");
-    console.log(this.idForUpdate);
-    console.log(employee);
-    this.employeeService.updateEmployee(this.idForUpdate,employee).subscribe(
-      (response: Employee) => {
-        console.log(response);
-        console.log("AICI0");
-        this.getEmployees();
-      
-      },
-      (error: HttpErrorResponse) => {
-        alert(error.message);
-      
-      }
-    );
+  open(content:any) {
+    this.modalService.open(content, {ariaLabelledBy: 'modal-basic-title'}).result.then((result) => {
+      this.closeResult = `Closed with: ${result}`;
+    }, (reason) => {
+      this.closeResult = `Dismissed ${this.getDismissReason(reason)}`;
+    });
   }
+  
+  private getDismissReason(reason: any): string {
+    if (reason === ModalDismissReasons.ESC) {
+      return 'by pressing ESC';
+    } else if (reason === ModalDismissReasons.BACKDROP_CLICK) {
+      return 'by clicking on a backdrop';
+    } else {
+      return `with: ${reason}`;
+    }
+  }
+
+
 
   public onAddEmloyee(addForm: NgForm): void {
     document.getElementById('add-employee-form')!.click();
@@ -138,8 +110,8 @@ export class EditEmployeeComponent implements OnInit {
     );
   }
   
-  public onDeleteEmloyee(employeeId: number ): void {
-
+  public onDeleteEmloyee(employeeId: number |undefined): void {
+if(employeeId)
     this.employeeService.deleteEmployee(employeeId).subscribe(
       (response: void) => {
         console.log(response);
@@ -150,6 +122,7 @@ export class EditEmployeeComponent implements OnInit {
       }
     );
   }
+
 
 
 }
