+++ src/app/service/employee.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from  '@angular/common/http'
 import { Observable } from 'rxjs';
-import { Employee } from './employee';
+import { Employee } from '../model/employee';
 import { environment } from 'src/environments/environment';
 
 @Injectable({
